Little Refactoring
Split the main file for clearer code
Improved security database
